backend/requirements.txt
fastapi
pymongo
bson
pydantic
google-generativeai